fix: check animation on the splash screen activity
