+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
       android:theme="@style/AppTheme.NoActionBar">
 
     </activity>
-    <activity android:name=".MainActivity" >
+    <activity android:name=".activities.MainActivity" >
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
         <category android:name="android.intent.category.LAUNCHER" />
